Update findLongestWordLength
If there's an empty string, then we have a case for that. Now, I'm going to specify a regular expression for actual sentences. Meaning, if there is a string that posses punctuation, then I will remove the punctuation to allow for accurate output.

As you can see, ' PRESTIDIGITATION ' only has 16 digits, yet 17 is printed out because of the punctuation.